refactor(desktop): extract the launcher payload lifecycle into updater/payload
Mechanical move out of updater.ts: payload manifest validation, archive
extraction (ditto/7z + mac xattr scrub), prepared-release verification,
payload activation with attempt pre-arming, and the launcher-side
cleanup machinery (deferred failures, retention, cold-start retry sweep,
manual-clear state reset) move to updater/payload.ts. updater.ts
re-exports LauncherPayloadExtractInput so the public surface is
unchanged; orphaned imports are pruned. No test files change: payload
coverage is integration-style and the desktop suite holds at 269.
